+++ src/bach/index.mjs
@@ -1,6 +1,6 @@
 const { series } = await import('./series.js');
-const { parallel } = await import('./parallel.js');
 const { settleSeries } = await import('./settleSeries.js');
-const { settleParallel } = await import('./settleParallel.js');
 
-export { series, parallel, settleSeries, settleParallel };
+export { series, settleSeries };
+export { parallel } from './parallel.mjs';
+export { settleParallel } from './settleParallel.mjs';
